+++ src/components/PatternScaling.jl
@@ -18,7 +18,11 @@
 
     function run_timestep(pp, vv, dd, tt)
         for cc in dd.country
-            vv.scale_country[tt, cc] = pp.ps_alpha[cc] + pp.ps_beta[cc] * log(pp.GMST_2010 + pp.T_AT[tt] - pp.T_AT_2010 - 18.825)
+            if pp.GMST_2010 + pp.T_AT[tt] - pp.T_AT_2010 - 18.825 > 0
+                vv.scale_country[tt, cc] = pp.ps_alpha[cc] + pp.ps_beta[cc] * log(pp.GMST_2010 + pp.T_AT[tt] - pp.T_AT_2010 - 18.825)
+            else
+                vv.scale_country[tt, cc] = pp.ps_alpha[cc] + pp.ps_beta[cc] * log(pp.GMST_2010 - 18.825)
+            end
             vv.T_country[tt, cc] = (pp.GMST_2010+pp.T_AT[tt]-pp.T_AT[TimestepIndex(1)])*vv.scale_country[tt, cc]
         end
     end
